+++ NEWS.md
@@ -1,6 +1,7 @@
 # tabr 0.4.2
 
 * Added utility functions for retrieving LilyPond version and installation directory and `tabr` LilyPond API details.
+* Fixed bug where multiple sharps in tuning broke tab staff string label LilyPond syntax.
 
 # tabr 0.4.1